+++ src/unounclass.cpp
@@ -1,13 +1,13 @@
 #include "main.hpp"
 
-#if (U_ICU_VERSION_MAJOR_NUM >= 71)
+#if (U_ICU_VERSION_MAJOR_NUM == 71)
 #include <unicode/unounclass.h>
-#endif // (U_ICU_VERSION_MAJOR_NUM >= 71)
+#endif // (U_ICU_VERSION_MAJOR_NUM == 71)
 
 using namespace icu;
 
 void init_unounclass(py::module &m) {
-#if (U_ICU_VERSION_MAJOR_NUM >= 71)
+#if (U_ICU_VERSION_MAJOR_NUM == 71)
   //
   // icu::NounClass
   //
@@ -22,5 +22,5 @@ void init_unounclass(py::module &m) {
       .value("PERSONAL", PERSONAL)
       .value("COMMON", COMMON)
       .export_values();
-#endif // (U_ICU_VERSION_MAJOR_NUM >= 71)
+#endif // (U_ICU_VERSION_MAJOR_NUM == 71)
 }